[mr2423] rewrite MapObjectsStore v0.13   - use GuidSet for make object-safe UpdateDataMapType (this fix at least pet action bar after mount)   - use unordered_map instead of TypeUnorderedMapContainer template collection   - full rewrite object update making system, make dependent update queue for players items   - add Transport objects to current Map storing/updating system (currently by half-hack)   - simplify code   - use storing guids instead of pointers in Camera/Viewpoint subsystem for fit new object storing system   - fix very old bug with Sentry totem (appears only now).
    NEED TEST and measurements CPU load and resource leaks! related to #1165, #1083, #852

+++ src/game/Unit.cpp
@@ -307,7 +307,7 @@ Unit::Unit() :
 Unit::~Unit()
 {
     if (IsInWorld())
-        Object::RemoveFromWorld();
+        WorldObject::RemoveFromWorld(true);
 
     ResetMap();
 
@@ -11212,11 +11212,11 @@ uint32 Unit::GetCreatePowers( Powers power ) const
 
 void Unit::AddToWorld()
 {
-    Object::AddToWorld();
+    WorldObject::AddToWorld();
     ScheduleAINotify(0);
 }
 
-void Unit::RemoveFromWorld()
+void Unit::RemoveFromWorld(bool remove)
 {
     // cleanup
     if (IsInWorld())
@@ -11232,14 +11232,17 @@ void Unit::RemoveFromWorld()
             MAPLOCK_WRITE(this,MAP_LOCK_TYPE_AURAS);
             CleanupDeletedHolders(true);
         }
-        GetViewPoint().Event_RemovedFromWorld();
     }
+    GetViewPoint().Event_RemovedFromWorld();
 
-    Object::RemoveFromWorld();
+    WorldObject::RemoveFromWorld(remove);
 }
 
 void Unit::CleanupsBeforeDelete()
 {
+    if (!IsInWorld())
+        return;
+
     if (m_uint32Values)                                      // only for fully created object
     {
         if (GetVehicle())
@@ -11248,11 +11251,11 @@ void Unit::CleanupsBeforeDelete()
             RemoveVehicleKit();
         InterruptNonMeleeSpells(true);
         KillAllEvents(false);                      // non-delatable (currently casted spells) will not deleted now but it will deleted at call in Map::RemoveAllObjectsInRemoveList
-        if (IsInWorld())
-            CombatStop();
+        CombatStop();
         ClearComboPointHolders();
-        DeleteThreatList();
-        if (GetTypeId()==TYPEID_PLAYER)
+        if (CanHaveThreatList())
+            DeleteThreatList();
+        if (GetTypeId() == TYPEID_PLAYER)
             getHostileRefManager().setOnlineOfflineState(false);
         else
             getHostileRefManager().deleteReferences();
